Rename plotting.py to util.py.
